+++ bambi/defaults/families.py
@@ -7,6 +7,7 @@ from bambi.families.univariate import (
     Binomial,
     Categorical,
     Cumulative,
+    Exponential,
     Gamma,
     Gaussian,
     HurdleGamma,
@@ -20,6 +21,7 @@ from bambi.families.univariate import (
     StudentT,
     VonMises,
     Wald,
+    Weibull,
     ZeroInflatedBinomial,
     ZeroInflatedNegativeBinomial,
     ZeroInflatedPoisson,
@@ -104,6 +106,15 @@ BUILTIN_FAMILIES = {
         },
         "link": {"a": "log"},
         "family": DirichletMultinomial,
+    },
+    "exponential": {
+        "likelihood": {
+            "name": "Exponential",
+            "params": ["mu"],
+            "parent": "mu",
+        },
+        "link": {"mu": "log"},
+        "family": Exponential,
     },
     "gamma": {
         "likelihood": {
@@ -242,6 +253,16 @@ BUILTIN_FAMILIES = {
         "link": {"mu": "inverse_squared", "lam": "log"},
         "family": Wald,
         "default_priors": {"lam": "HalfCauchy"},
+    },
+    "weibull": {
+        "likelihood": {
+            "name": "Weibull",
+            "params": ["mu", "alpha"],
+            "parent": "mu",
+        },
+        "link": {"mu": "log", "alpha": "log"},
+        "family": Weibull,
+        "default_priors": {"alpha": "HalfCauchy"},
     },
     "zero_inflated_binomial": {
         "likelihood": {