+++ docs/CHANGELOG.md
@@ -1,14 +1,16 @@
 # Changelog
 
-
 ## 0.X.X
 
 ### New features
 
+* Bambi now supports censored responses (#697)
+* Implement `"exponential"` and `"weibull"` families (#697)
+* Add `"kidney"` dataset (#697)
+
 ### Maintenance and fixes
 
 ### Documentation
-
 
 ### Deprecation